Securing tasks and assigning to logged user

+++ backend/repository/taskRepository.go
@@ -15,20 +15,11 @@ func NewTaskRepository(db *sql.DB) TaskRepository {
 	}
 }
 
-func (r *TaskRepository) GetTasks(searchTerm string) (*[]entitites.Task, error) {
-	query := "SELECT id, title, is_completed FROM tasks"
-	if searchTerm != "" {
-		query += " WHERE ts @@ phraseto_tsquery('english', $1)"
-	}
+func (r *TaskRepository) GetTasksByUser(user *entitites.User) (*[]entitites.Task, error) {
+	query := "SELECT id, title, is_completed FROM tasks WHERE user_id = $1"
 	query += " ORDER BY id ASC"
 
-	var err error
-	var rows *sql.Rows
-	if searchTerm != "" {
-		rows, err = r.db.Query(query, searchTerm)
-	} else {
-		rows, err = r.db.Query(query)
-  }
+  rows, err := r.db.Query(query, user.Id)
 
 	if err != nil {
 		return nil, err
@@ -54,13 +45,13 @@ func (r *TaskRepository) GetTasks(searchTerm string) (*[]entitites.Task, error) 
 	return &tasks, nil
 }
 
-func (r *TaskRepository) CreateTask(title string) (*entitites.Task, error) {
+func (r *TaskRepository) CreateTask(title string, user *entitites.User) (*entitites.Task, error) {
 	var task entitites.Task
 	task.IsCompleted = false
 	task.Title = title
 
-	query := "INSERT INTO tasks (title) VALUES ($1) RETURNING id"
-	err := r.db.QueryRow(query, title).Scan(&task.Id)
+	query := "INSERT INTO tasks (title, user_id) VALUES ($1, $2) RETURNING id"
+	err := r.db.QueryRow(query, title, user.Id).Scan(&task.Id)
 	if err != nil {
 		return nil, err
 	}
@@ -68,15 +59,15 @@ func (r *TaskRepository) CreateTask(title string) (*entitites.Task, error) {
 	return &task, nil
 }
 
-func (r *TaskRepository) UpdateTask(id int, title string, isCompleted bool) (int, error) {
+func (r *TaskRepository) UpdateTask(id int, title string, isCompleted bool, user *entitites.User) (int, error) {
 	isCompletedBit := 0
 	if isCompleted {
 		isCompletedBit = 1
 	}
 
-	query := "UPDATE tasks SET title = $1, is_completed = $2 WHERE id = $3"
+	query := "UPDATE tasks SET title = $1, is_completed = $2 WHERE id = $3 AND user_id = $4"
 
-	result, err := r.db.Exec(query, title, isCompletedBit, id)
+	result, err := r.db.Exec(query, title, isCompletedBit, id, user.Id)
 	if err != nil {
 		return -1, err
 	}
@@ -87,9 +78,9 @@ func (r *TaskRepository) UpdateTask(id int, title string, isCompleted bool) (int
 	return int(affected), nil
 }
 
-func (r *TaskRepository) DeleteTask(id int) (int, error) {
-	query := "DELETE FROM tasks WHERE id = $1"
-	result, err := r.db.Exec(query, id)
+func (r *TaskRepository) DeleteTask(id int, user *entitites.User) (int, error) {
+	query := "DELETE FROM tasks WHERE id = $1 AND user_id = $2"
+	result, err := r.db.Exec(query, id, user.Id)
 	if err != nil {
 		return -1, err
 	}
